2
Completed the main functionality of the questions:
When the user selects start the first question (from level 2) shows up.
If it is answered correctly then the user gets 2 points and the next questions is on level 3.
Else the next question is on level 1.

There are 5 rounds. 
GENERALLY:
If the user answers correctly a level 1 question  he gets one point and the next question is on level 2.
Else he gets 0 points and the next question is on level 1.

If the user answers correctly a level 2 question he gets 2 points and the next questions is on level 3.
Else the next question is on level 1.

If the user answers correctly a level 3 question he gets 3 points and the next questions is on level 3.
Else the next question is on level 2.

The score of the user is temporarily saved in a text file. 
Also, are used additional text files in order to remember the current level, how many questions are answered so far and the number of the asking question.